Implemented Back-End for Adding Items to Cart
Implemented a JS backend so that items can be added to the cart as objects with name, price, and quantity values. The amount of items added to the cart will pop up on the top-right of the cart icon as a bubble. Next thing to do is to implement the check-out page so that the items that are added to the cart can be displayed properly for a check-out system.