Cart Info Saved in localStorage
Implemented a way of saving cart items and the number of cart items in the local storage so that the user's cart information doesn't reset after refreshing the page.

The user will get a pop-up notification 5 minutes after adding their last item to the cart that they haven't checked out in 5 minutes, and that their cart information will disappear after 25 minutes.

30 minutes after the user has added their last item to the cart, their cart information will be erased for not checking out on time.